Move tests to hits_test
